Add permission check messages

+++ langwatch/src/pages/[project]/datasets.tsx
@@ -31,17 +31,26 @@ import { Link } from "../../components/ui/link";
 import { Menu } from "../../components/ui/menu";
 import { toaster } from "../../components/ui/toaster";
 import { PageLayout } from "~/components/ui/layouts/PageLayout";
+import { PermissionAlert } from "../../components/PermissionAlert";
 
 export default function Datasets() {
   const addEditDatasetDrawer = useDisclosure();
   const uploadCSVModal = useDisclosure();
-  const { project } = useOrganizationTeamProject();
+  const { project, hasPermission } = useOrganizationTeamProject();
   const router = useRouter();
   const { openDrawer } = useDrawer();
+  const hasDatasetsViewPermission = hasPermission("datasets:view");
+  if (!hasDatasetsViewPermission) {
+    return (
+      <DashboardLayout>
+        <PermissionAlert permission="datasets:view" />
+      </DashboardLayout>
+    );
+  }
 
   const datasets = api.dataset.getAll.useQuery(
     { projectId: project?.id ?? "" },
-    { enabled: !!project }
+    { enabled: !!project },
   );
 
   const datasetDelete = api.dataset.deleteById.useMutation();
@@ -90,7 +99,7 @@ export default function Datasets() {
                           });
                           addEditDatasetDrawer.onClose();
                         },
-                      }
+                      },
                     );
                   }}
                 >
@@ -118,7 +127,7 @@ export default function Datasets() {
             },
           });
         },
-      }
+      },
     );
   };
